Added iotcrafter build (condensed all previous work into one commit as initial)

+++ scripts/chroot.sh
@@ -606,8 +606,10 @@ cat > "${DIR}/chroot_script.sh" <<-__EOF__
 			rm -f /tmp/repos.rcn-ee.net-archive-keyring.asc || true
 		fi
 		if [ "x${repo_external}" = "xenable" ] ; then
+		    if [ ! "x${repo_external_key}" = "x" ]; then
 			apt-key add /tmp/${repo_external_key}
 			rm -f /tmp/${repo_external_key} || true
+		    fi
 		fi
 		if [ "x${repo_flat}" = "xenable" ] ; then
 			apt-key add /tmp/${repo_flat_key}
